requirements.txt
Flask==3.0.3
Flask_SQLAlchemy==3.1.1
Flask_Migrate==4.0.5
openai>=1.0.0
pytest>=7.0.0
pytest-mock>=3.0.0